add delivery and statistics controller

+++ resources/views/front/wishlist.blade.php
@@ -191,18 +191,36 @@
                 e.preventDefault();
                 var id = $(this).data("id");
                 var token = $("meta[name='csrf-token']").attr("content");
-                $.ajax(
-                {
-                    url: "/wishlist/"+id,
-                    type: 'DELETE',
-                    data: {
-                        "id": id,
-                        "_token": token,
-                    },
-                    success: function (response){
-                        $(`#${id}`).remove();
-                    }
-                });
+                Swal.fire({
+                    title: 'Are you sure?',
+                    text: "You won't be able to revert this!",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Yes, delete it!'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        Swal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        );
+                        $.ajax({
+                            url: "/wishlist/"+id,
+                            type: 'DELETE',
+                            data: {
+                                "id": id,
+                                "_token": token,
+                            },
+                            success: function (response){
+                                $('span#count').empty();
+                                $('span#count').append(result);
+                                $(`#${id}`).remove();
+                            }
+                        });
+                    }})
+               
                 
             });
         });
